added lib folder for server side js files
